+++ modules/azurerm_subnet/examples/delegation/.terraform/modules/subnet_batch/modules/azurerm_virtual_network/examples/private-endpoint/main.tf
@@ -51,7 +51,7 @@ resource "azurerm_private_dns_zone" "blob" {
 
 # Virtual Network optimized for private endpoint scenarios
 module "virtual_network" {
-  source = "github.com/PatrykIti/azurerm-terraform-modules//modules/azurerm_subnet?ref=SNv1.0.1"
+  source = "github.com/PatrykIti/azurerm-terraform-modules//modules/azurerm_subnet?ref=SNv1.0.2"
 
   name                = "vnet-private-endpoint-example"
   resource_group_name = azurerm_resource_group.example.name