+++ modules/azurerm_subnet/examples/delegation/.terraform/modules/subnet_container_instances/modules/azurerm_storage_account/tests/fixtures/network/main.tf
@@ -45,7 +45,7 @@ resource "azurerm_subnet" "app" {
 }
 
 module "storage_account" {
-  source = "github.com/PatrykIti/azurerm-terraform-modules//modules/azurerm_subnet?ref=SNv1.0.1"
+  source = "github.com/PatrykIti/azurerm-terraform-modules//modules/azurerm_subnet?ref=SNv1.0.2"
 
   name                     = "dpcnet${var.random_suffix}"
   resource_group_name      = azurerm_resource_group.test.name