+++ modules/azurerm_subnet/examples/delegation/.terraform/modules/subnet_regular/modules/azurerm_virtual_network/tests/fixtures/negative/empty_address_space/main.tf
@@ -24,7 +24,7 @@ resource "azurerm_resource_group" "test" {
 }
 
 module "virtual_network" {
-  source = "github.com/PatrykIti/azurerm-terraform-modules//modules/azurerm_subnet?ref=SNv1.0.1"
+  source = "github.com/PatrykIti/azurerm-terraform-modules//modules/azurerm_subnet?ref=SNv1.0.2"
 
   # Invalid: empty address space
   name                = "vnet-empty-address"