+++ modules/azurerm_kubernetes_cluster/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+All notable changes to the Kubernetes Cluster Terraform module will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [1.0.3](https://github.com/PatrykIti/azurerm-terraform-modules/compare/AKSv1.0.2...AKSv1.0.3) (2025-09-22)
+
+### 🐛 Bug Fixes
+
+* **kubernetes-cluster,network-security-group,route-table,storage-account,subnet,virtual-network:** complete module implementation with tests and examples ([#67](https://github.com/PatrykIti/azurerm-terraform-modules/issues/67)) ([ddff590](https://github.com/PatrykIti/azurerm-terraform-modules/commit/ddff59038ed839e6ec5d0c538764c2b6fa56204c))
+* **kubernetes-cluster,network-security-group,route-table,storage-account,subnet,virtual-network:** complete module implementation with tests and examples ([#68](https://github.com/PatrykIti/azurerm-terraform-modules/issues/68)) ([51e7c3e](https://github.com/PatrykIti/azurerm-terraform-modules/commit/51e7c3e7c3226b16fe499579a65a5d585f4752b3))
+
+# Changelog
+
 All notable changes to this module will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),